// frontend/src/api/bindings/PlanetAttachment.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A planet attachment.
 */
export type PlanetAttachment = "DemilitarizedZone" | "DysonSphere" | "ParadiseWorld" | "TombOfEmphidia" | "BioticResearchFacility" | "CyberneticResearchFacility" | "PropulsionResearchFacility" | "WarfareResearchFacility" | "BioticResearchFacilityResources" | "CyberneticResearchFacilityResources" | "PropulsionResearchFacilityResources" | "WarfareResearchFacilityResources" | "LasaxSurvivors" | "MiningWorld" | "RichWorld" | "UITheProgenitor" | "Terraform" | "NanoForge";